# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{slug}/invitations": {
      "post": {
        "description": "Только администратор проекта (US-20). Роль администратора через приглашение выдать нельзя (D-05), срок жизни — 1, 7 или 30 дней (по умолчанию 7); бессрочных приглашений нет. По одной ссылке может вступить любое число людей, пока она действует. Полный адрес ссылки — в поле `url`.",
        "operationId": "ProjectInvitationsController_create",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "sladkiy-limit",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "403": {
            "description": "Не администратор проекта (`project_forbidden`)"
          },
          "404": {
            "description": "Проекта нет или пользователь не участник"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Создать ссылку-приглашение",
        "tags": [
          "invitations"
        ]
      },
      "get": {
        "description": "Все приглашения проекта, сначала новые: действующие, истёкшие и отозванные с числом вступивших по каждому (US-22).",
        "operationId": "ProjectInvitationsController_list",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор следующей страницы из поля `nextCursor`",
            "schema": {
              "minLength": 1,
              "maxLength": 256,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationListDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный курсор (`invalid_cursor`)"
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "403": {
            "description": "Не администратор проекта (`project_forbidden`)"
          },
          "404": {
            "description": "Проекта нет или пользователь не участник"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Приглашения проекта",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/projects/{slug}/invitations/{id}/revoke": {
      "post": {
        "description": "Отозвать может любой администратор проекта, в том числе не тот, кто создавал ссылку. Сразу после отзыва переход по ссылке даёт 410. Уже вступившие остаются участниками (US-22).",
        "operationId": "ProjectInvitationsController_revoke",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "403": {
            "description": "Не администратор проекта (`project_forbidden`)"
          },
          "404": {
            "description": "Приглашения нет (`invitation_not_found`)"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Отозвать приглашение",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/invitations/{token}": {
      "get": {
        "description": "Данные для экрана подтверждения: название проекта, обложка и роль, которую получит человек. Ни задач, ни очередей, ни участников (US-21). Неизвестный токен — 404, истёкший или отозванный — 410; по сообщению нельзя понять, существует ли такой проект. `alreadyMember: true` означает, что экран показывать не нужно — клиент сразу открывает проект.",
        "operationId": "InvitationAcceptController_preview",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Токен из ссылки `/invite/<token>`",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPreviewDto"
                }
              }
            }
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "404": {
            "description": "Токен неизвестен (`invitation_not_found`)"
          },
          "410": {
            "description": "Истекло или отозвано (`invitation_inactive`)"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Приглашение по токену",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/invitations/{token}/accept": {
      "post": {
        "description": "Человек становится участником с ролью из приглашения, а его email попадает в список доступа с источником `invitation` — со следующего входа ссылка ему уже не нужна (ADR-0006, п. 3). Тот, кто уже состоит в проекте, просто получает адрес проекта: роль не меняется и не понижается (US-21).",
        "operationId": "InvitationAcceptController_accept",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptInvitationResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Нет сессии или она погашена"
          },
          "404": {
            "description": "Токен неизвестен (`invitation_not_found`)"
          },
          "410": {
            "description": "Истекло или отозвано (`invitation_inactive`)"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Вступить в проект по приглашению",
        "tags": [
          "invitations"
        ]
      }
    },
...
  },
  "info": {
    "title": "SL Tracker API",
    "description": "HTTP API трекера задач SL Tracker. Единственный источник правды по контракту: файл генерируется из кода, руками не редактируется.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sl_session",
        "description": "Сессия веб-клиента: httpOnly Secure SameSite=Lax cookie (ADR-0001)."
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "Сессия будущего мобильного клиента (ADR-0002)."
      }
    },
    "schemas": {
...
      "CreateInvitationDto": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "member",
              "reader"
            ],
            "type": "string",
            "description": "Роль для тех, кто вступит. Роли «администратор» в списке нет (D-05)."
          },
          "expiresInDays": {
            "enum": [
              1,
              7,
              30
            ],
            "type": "number",
            "default": 7,
            "description": "Срок жизни ссылки в днях. Бессрочных приглашений нет (US-20)."
          }
        },
        "required": [
          "role"
        ]
      },
      "InvitationAuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "nullable": true,
            "example": "Анна Иванова"
          }
        },
        "required": [
          "id",
          "displayName"
        ]
      },
      "InvitationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "enum": [
              "member",
              "reader"
            ],
            "type": "string",
            "description": "Роль, которую получит вступивший. Администратора выдать нельзя (D-05)."
          },
          "state": {
            "enum": [
              "active",
              "expired",
              "revoked"
            ],
            "type": "string",
            "description": "Вычисляется из срока и отметки об отзыве, отдельно не хранится. Истёкшие и отозванные остаются в списке, но повторно активировать их нельзя (US-22)."
          },
          "url": {
            "type": "string",
            "nullable": true,
            "description": "Полная ссылка-приглашение. Заполнена только у действующего приглашения: у истёкшего и отозванного её нет и копировать нечего (US-22)."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "revokedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "description": "Кто создал приглашение",
            "allOf": [
              {
                "$ref": "#/components/schemas/InvitationAuthorDto"
              }
            ]
          },
          "acceptedCount": {
            "type": "number",
            "description": "Сколько человек вступило по этой ссылке (US-22)"
          }
        },
        "required": [
          "id",
          "role",
          "state",
          "url",
          "expiresAt",
          "revokedAt",
          "createdAt",
          "createdBy",
          "acceptedCount"
        ]
      },
      "InvitationListDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Сначала новые",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitationDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "total": {
            "type": "number",
            "description": "Всего приглашений у проекта, включая истёкшие и отозванные"
          }
        },
        "required": [
          "items",
          "nextCursor",
          "total"
        ]
      },
      "InvitationPreviewDto": {
        "type": "object",
        "properties": {
          "projectName": {
            "type": "string",
            "example": "Сладкий лимит"
          },
          "projectSlug": {
            "type": "string",
            "example": "sladkiy-limit"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true,
            "description": "Подписанная ссылка на обложку"
          },
          "role": {
            "enum": [
              "admin",
              "member",
              "reader"
            ],
            "type": "string",
            "description": "Роль, которую человек получит. Если он уже участник — его текущая роль: приглашение её не меняет и не понижает (US-21)."
          },
          "alreadyMember": {
            "type": "boolean",
            "description": "Пользователь уже состоит в проекте: экран подтверждения не показывается, клиент сразу открывает проект (US-21)."
          }
        },
        "required": [
          "projectName",
          "projectSlug",
          "coverUrl",
          "role",
          "alreadyMember"
        ]
      },
      "AcceptInvitationResultDto": {
        "type": "object",
        "properties": {
          "projectSlug": {
            "type": "string",
            "example": "sladkiy-limit",
            "description": "Куда переходить после вступления"
          },
          "role": {
            "enum": [
              "admin",
              "member",
              "reader"
            ],
            "type": "string",
            "description": "Роль пользователя в проекте после приёма"
          },
          "alreadyMember": {
            "type": "boolean",
            "description": "Пользователь уже был участником; роль не изменилась"
          }
        },
        "required": [
          "projectSlug",
          "role",
          "alreadyMember"
        ]
      },
...
    }
  }
}